Add server-side active close mechanism
This patch set fixes the behavior of Trojan not immediately attempting to close a frontend connection if the Trojan backend unexpectedly closed the connection.

In some cases, an attacker can trick the backend into closing the connection immediately after sending valid data with carefully constructed contents. For connections that are not relayed by Trojan, the attack can observe the closure of the connection immediately. However, in scenarios where Trojan is present, the connection will be kept even after the attacker receives the aforementioned data. This difference in behavior can be used to detect the presence of Trojan.

The behavioral difference does not result in the leak of confidential user data.